added logic to save favorite, now need to list the favorites on main activity

+++ app/src/main/java/com/keanesf/popmovies/data/database/FavoriteEntry.java
@@ -17,18 +17,18 @@ import java.util.Date;
 public class FavoriteEntry {
 
     @PrimaryKey
-    private int id;
+    private Long id;
 
     // Constructor used by Room to create FavoriteEntries
-    public FavoriteEntry(int id) {
+    public FavoriteEntry(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
